+++ zerorle.h
@@ -1,4 +1,4 @@
 #pragma once
 
-int deRLE(const BYTE * in, BYTE * out, const int length, const int level);
-unsigned int TestAndRLE(BYTE * const __restrict in, BYTE * const __restrict out1, BYTE * const __restrict out3, const unsigned int length, int * level);
+size_t deRLE(const BYTE* in, BYTE* out, const size_t length, const BYTE level);
+size_t TestAndRLE(BYTE* const __restrict in, BYTE* const __restrict out1, BYTE* const __restrict out3, const size_t length,  int* level);
